updated namespaces and removed some unused classes
