--- modules/covid_vaccine/app/serializers/covid_vaccine/registration_submission_serializer.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-module CovidVaccine
-  class RegistrationSubmissionSerializer < ActiveModel::Serializer
-    attribute :created_at
-
-    def id
-      object.sid
-    end
-  end
-end
