+++ modules/covid_vaccine/app/serializers/covid_vaccine/v0/registration_submission_serializer.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module CovidVaccine
+  module V0
+    class RegistrationSubmissionSerializer < ActiveModel::Serializer
+      attribute :created_at
+
+      def id
+        object.sid
+      end
+    end
+  end
+end
