+++ modules/covid_vaccine/spec/request/covid_vaccine/v0/registration_request_spec.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+require 'rails_helper'
+
+RSpec.describe 'Covid Vaccine Registration', type: :request do
+  include SchemaMatchers
+
+  context 'with an unauthenticated user' do
+  end
+
+  context 'with a loa1 user' do
+  end
+
+  context 'with a loa3 user' do
+  end
+end
